Add files via upload
The discriminative model reproduces the Maxwell-Juttner distribution by first passing data through a classifier that accepts data if the criteria in the Sobol algorithm is met, the accepted data is then passed into a regression network to estimate the resulting MJ value. This model mimics the Sobol algorithm in two steps.