Add files via upload
We implement a vanilla GAN capable of sampling from a skewed Gaussian distribution with skew parameter of 7.  To do this, we introduce a final layer SiLU activation function on our generator which works well. We also sample towards the end of training (we also reduce our learning rates) to introduce randomness into our sampled numbers.